Rename enroll to accessTokens. Remove auth and enroll view
