+++ app.js
@@ -6,8 +6,7 @@ var logger = require("morgan");
 var sassMiddleware = require("node-sass-middleware");
 
 var indexRouter = require("./routes/index");
-var enrollRouter = require("./routes/enroll");
-var authRouter = require("./routes/auth");
+var accessTokensRouter = require("./routes/accessTokens");
 var challengeRouter = require("./routes/challenge");
 var webhookRouter = require("./routes/webhook");
 
@@ -36,8 +35,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/", indexRouter);
-app.use("/enroll", enrollRouter);
-app.use("/auth", authRouter);
+app.use("/accessTokens", accessTokensRouter);
 app.use("/challenge", challengeRouter);
 app.use("/webhook", webhookRouter);
