+++ views/enrollment.pug
@@ -15,14 +15,6 @@ block content
         pre
           = uri
 
-      if jwt
-        hr.my-4
-        h6 JWT Header
-        pre
-          = JSON.stringify(jwt.header, null, 4)
-        h6 JWT Payload
-        pre
-          = JSON.stringify(jwt.payload, null, 4)
     .col-md-6
       if qr
         img(src=qr).img-fluid.mx-auto.img-thumbnail
