+++ views/index.pug
@@ -8,13 +8,13 @@ block content
             a(href="/challenge") Create Push Challenge: 
             span Helps to create a Factor Push Challenge with custom details.
         li
-            b /enroll 
-            span A POST route to generate a JWT with the required payload to create a factor with the Verify API
+            b /accessTokens 
+            span A POST route to generate an access token to create a factor with the Verify API
 
-    h4 Enrollment JWT Endpoint
+    h4 Access Tokens Endpoint
     div.mt-4
         span.badge.badge-success.mr-3 POST
-        code /enroll
+        code /accessTokens
     div.mt-4
         h5 Params
         dl
