+++ views/layout.pug
@@ -12,8 +12,6 @@ html
 
   body
     +navbar_contained("Twilio", "navBar", "dark", "blue", "/")
-      +nav_item("/enrollment")
-        = "Enrollment"
       +nav_item("/challenge")
         = "Challenge"
